refactor: extract select_model to shared utility
Extracted the duplicated `select_model` function from multiple Streamlit
application files (`chat.py`, `youtube_summarizer.py`, and
`website_summarizer.py`) into a common utility module `src/ai_agent/utils.py`.

Standardized the list of available models and improved module importability
by guarding external dependencies with a `try-except` block in the utility
module, ensuring compatibility with existing tests.
