+++ src/ai_agent/streamlit/website_summarizer.py
@@ -4,14 +4,9 @@ from langchain_core.prompts import ChatPromptTemplate
 from langchain_core.output_parsers import StrOutputParser
 from urllib.parse import urlparse, urlunparse
 
-# models
-from langchain_openai import ChatOpenAI
-from langchain_anthropic import ChatAnthropic
-from langchain_google_genai import ChatGoogleGenerativeAI
-
 import requests
 from bs4 import BeautifulSoup
-from ai_agent.utils import validate_url
+from ai_agent.utils import validate_url, select_model
 
 MAX_CONTENT_LENGTH = 20000
 
@@ -31,44 +26,6 @@ def init_page():
     )
     st.title("Website Summarizer📝")
     st.sidebar.title("Options")
-
-
-def select_model():
-    """利用するLLMをサイドバーの選択状態によって切り替える"""
-    temperature = st.sidebar.slider(
-        "Temperature", min_value=0.0, max_value=1.0, value=0.0, step=0.1
-    )
-    models = (
-        "Open AI GPT-3.5-turbo",
-        "Open AI GPT-4o",
-        "Claude 3.5 Haiku",
-        "Google Gemini 1.5 Flash",
-    )
-    model = st.sidebar.radio("Choose a model:", models)
-
-    match model:
-        case "Open AI GPT-3.5-turbo":
-            st.session_state.model_name = "gpt-3.5-turbo"
-            return ChatOpenAI(
-                model=st.session_state.model_name, temperature=temperature
-            )
-        case "Open AI GPT-4o":
-            st.session_state.model_name = "gpt-4o"
-            return ChatOpenAI(
-                model=st.session_state.model_name, temperature=temperature
-            )
-        case "Claude 3.5 Haiku":
-            st.session_state.model_name = "claude-3-5-haiku-20241022"
-            return ChatAnthropic(
-                model=st.session_state.model_name, temperature=temperature
-            )
-        case "Google Gemini 1.5 Flash":
-            st.session_state.model_name = "gemini-1.5-flash-latest"
-            return ChatGoogleGenerativeAI(
-                model=st.session_state.model_name, temperature=temperature
-            )
-        case _:
-            raise ValueError("Invalid model selected.")
 
 
 def init_chain():
